Replace metrics-prometheus-exporter http server with adapter.
Metrics-prometheus-exporter::install() creates an HTTP server that
listens on port 9100. Instead, we can use the HTTP server that the
adapter already runs to service prometheus scrape requests. This
is accomplished by installing a global metrics recorder, as the
metrics-prometheus-exporter did previously, and storing a handle to
it, used to retrieve the scrape as a string when a request is made
to /prometheus.

The port in the vector config for scraping prometheus metrics
is updated accordingly to 6034, the adapter http server.

Change-Id: Ifac7f3e7fd72797e9c3e93c2a48ba106a8b43d0a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/923
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>