replicators: Log when mysql snapshot is finished
Similar to what we do for postgresql, log a "snapshot finished" log
message once the initial mysql snapshot has finished and we've obtained
the binlog position. We're going to be referring to this log line in the
setup docs directly, so we want it to be consistent between the two
upstreams.

Change-Id: I1bc1713d529263c4f8d31768102cd818828795f3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/929
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>