images: Tag volumes with the deployment name
Rather than just tagging all readyset server volumes with the same
key/value pair, use the name of the readyset deployment as the tag
value, so that multiple deployments in the same aws account don't try to
mount each other's volumes. This might not matter *right now*, but feels
relevant for later.

Change-Id: Iaceafbdb1194522bb4a737bb30fa85ddb587bb5a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/937
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>