Move readyset-server config inside ensure-ebs-volume
ensure-ebs-volume is going to move to become a daemon that runs in the
background soon - as such we won't be able to wait for it to
successfully mount the volume in the user-data-init.sh script, so that
has to happen inside ensure-ebs-volume itself. This threads through the
volume ID of the ebs volume that's mounted (which is just the EBS volume
ID itself now, not a random UUID, to make it easier to correlate with
state visible through the AWS console), then writes that to the
/etc/default EnvironmentFile for the readyset-server service itself,
then does the actual starting of the readyset-server (and signalling
health to cloudformation) as an exec'd
process within ensure-ebs-volume.

Change-Id: Id131af6675a3ea7a5fa60a1e7c23b11cd85decee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/939
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>