images: Run ensure-ebs-volume as a systemd unit
Rather than running ensure-ebs-volume in the foreground as part of the
user-data-init script, add a systemd unit for it configured via an
EnvironmentFile that gets enabled and started by the user-data-init
script. This currently just exits after successfully mounting the volume
and starting readyset-server, but in the future will be modified to
listen to an SNS topic for ASG scaledown events to unmount and detach
the volume gracefully before the instance is terminated.

Change-Id: I2fa8b8fca059547420c35e23e9ca8100a8c6838b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/940
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>