ops/cfn: Update policy for new volume tagging
6e1799791 (images: Tag volumes with the deployment name, 2021-10-28)
changed the behavior of ensure-ebs-volume to tag images with the name of
the deployment rather than just "true", but didn't update the policy for
the instance to allow mounting volumes with this name.

Perhaps at some point in the future we'll make this conditional to
specifically the name of the deployment, but for now we'll just make
this allow any volume where this tag is present.

Change-Id: I5bcd6b56af221b18507c52d9d312d5f4482465c8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/942
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>