ensure-ebs-volume: Listen for terminating events
As part of ensure-ebs-volume, listen on a (configurable) SQS queue for
instance lifecycle hook events for the instance we're running for being
terminated by the ASG, and in response kill noria-server, unmount the
disk, and detach the ebs volume from the instance, then send the
"CompleteLifecycleAction" API request to signal the ASG that we're
ready to continue terminating.

This ended up being *quite* a bit of code - notably some stuff has been
refactored to store state as fields on Opts rather than passing them
around in return values and arguments, mostly because I don't want to go
down the path of returning enormous opaque tuples from things.

This also modifies the CFN template for this queue to allow the instance
role to do the requisite API calls to read from the queue and complete
the lifecycle action.

Change-Id: Ia7ce2674bf09a5cbd940b0e73ea60f17fa0ca9d0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/944
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>