ops/cfn: Set VisibilityTimeout=0 for queue
This queue is going to contain termination messages for all instances in
the ASG, and then each instance is going to ignore all messages that
aren't for its own instance id. We don't want one instance receiving a
message not for its own instance id to prevent other instances from
receiving that message for a few seconds, which is the default behavior
if this value isn't set to 0.

Change-Id: I4dfb520c08fafe60180acd7d84fcbd8bdcff12c8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/946
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>