Adds registration and retrieval of adapter addrs
This commit adds the functionality for registering an adapters http
endpoint with consul, and the ability to retrieve all adapter addresses
back out of consul.

This is currently set to todo! for other authorities as we only need
consul for now. Also includes a new test to prove out this
functionality, which has been confirmed as passing with a local
deployment of consul.

Change-Id: Idb44754791367b9bcc6f3790113c8d0c0457320f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/947
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>