Discover adapter endpoints in metrics aggregator
This commit adds logic for the metrics aggregator to retrieve the up to
date list of current adapter endpoints from consul on each
reconciliation pass. This also removes the older flag for manually
passing in a list of adapter addresses, and instead requires passing in
an address for consul (where we grab the current list of adapter
addresses from).

This commit also updates the docker compose file used for local testing
to be setup for this new method of service discovery.

Tested locally and working correctly.

Change-Id: I664244531fe8ec1d489bbaea00d4a2e2fc748ab5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/949
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>