ensure-ebs-volume: Fetch all EBS volumes
Instead of filtering server side on EBS volumes, we now need to see all
the volumes during the initial query to find which one to attach.

This attempts to keep the same behavior as the previous version just
moving the filtering to the other side. There is now a place as well
to put the new logic.

Change-Id: Ib686fffbc2df0003fedd1a0061ad7404c3dbebfb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/954
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>