ops/cfn: Enable EBS Optimization for ReadySet Server EC2s
To minimize contention in network bandwidth on our ReadySet Server
instances, we should enable EBS optimization. The benefit to doing so
is that interactions between EC2 <-> EBS will no longer be included in
the main ENI for the instance, and instead, will be on a separate ENI.

This will ensure that ReadySet can continue to communicate with other
components in the stack without running into network throughput
bottlenecks due to disk writes that may be caused by: RocksDB,
swap, or other consumers of disk i/o. When attaching multiple EBS
volumes to a single instance, this setting becomes even more pertinent.

Also as part of this CL, moving away from t3 instance types in favor of
instances that are less prone to CPU steal. Monitoring instance
is now set be a c5.large, and the adapter instances are proposed as
m5.large.

Adding constraint description and parameter validation for instance
types to ensure they're within the 5th generation.

Change-Id: I542a9df27af5be7309fc951cec2e9a5905e0ae1a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/956
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>