ensure-ebs-volume:  Implement instance termination piece of the workflow
When no volume is found, and we can't create one because we've already
reached our volume limit, grab a list of running instances in our
availability zone.  If all of them are up and healthy, terminate
ourselves so we can come back up in another AZ; if any are shutting
down, wait for them to shut down.

Change-Id: I24c94365a3d0f38131a1b85d21b2135a09ada897
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/958
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>