Packer for metrics-aggregator
This adds a new provisioner `metrics-aggregator` which has the scripts
and files required to run an instance of the metrics aggregator service.

The metrics-aggregator runs as a systemd service. It uses consul to
discover adapters in the system, pull their respective allow and deny
lists, aggregating all of that data together. It then queries prometheus
server for per query latency metrics and bundles that together with the
aggregated allow list.

Change-Id: I4a43162d036fe0954bf97ee3508b9071bcab60b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/959
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>