noria-client:  Reduce memory "leaked" per query with QCA enabled
Using the term "leaked" pretty loosely, here.  It's intentional (and
desirable!) to store a copy of the parsed query with each
QueryExecutionEvent, and that data is being utilized.  What we're doing
here is reducing the amount of data copied for each query when QCA is
enabled.

As part of that process, the original QueryCoverageInfo scheme was removed.
That was already on the docket for removal, having been superseded by Live QCA,
and it simplified the implementation of this to remove it rather than optimizing
it.  Backend::mirror_execute() had to be modified fairly substantially to handle
that removal.

Other than that, changes included low-hanging fruit in terms of eliminating
extraneous clones that were happening with QCA enabled.

Change-Id: Ic533bee85a3982c939b94c40756431eec93d29a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/989
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>