eviction: Account for keys in reader's memory size.
This adds the addition of keys to a reader's memory size, which
begins tracking filled, but empty entries in the reader. These
can grow extremely large on sparse key sets when reads lead to
keys that are not in the base table.

Accounting for this allows us to more accurately track memory
utilization and trigger eviction when appropriate.

Change-Id: Ic4de6af192dc8e0257a54ef45b61be4911fda8ec
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/973
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>