eviction: Estimate the size of the batch to evict in the reader.
Prior to this CL, we always evict 16 keys at a time during eviction
at reader nodes. However, the random eviction operation called
on the reader_map is expensive and publishing updates of 16
random evictions at a time takes upwards of 20 seconds locally
if the amount of data to evict is 1MB. This is problematic
as domains are singled threaded and processing an eviction
blocks queries that need replay's handled.

This CL adds:
  * The size of each key value pair when it doesn't have values
    into the memory tracking.
  * Using the size of key value pairs as an estimate for how many
    keys to evict. This will evict more than required as the
    estimate is a lower bound.

Change-Id: I0d8bca5495fac7e05e41d4215715f914b4cbdef4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/974
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>