eviction: +State tracking and large batches for MemoryState eviction.
This adds tracking the amount of memory stored in the key and
Row::default to the memory state. Previously we only included
the memory of rows that were added to the state, but with
requests that don't map to a row in the base table, we still
add state.

Additionally, we make similar changes to batch evictions into a single
large batch for MemoryState, as made for reader maps:
https://gerrit.readyset.name/c/readyset/+/974

Change-Id: I8c4635c7ca6999c622dfa53422f16585194f0df6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/977
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>