eviction: Only call `evict_bytes` once per node.
Eviction at dataflow nodes is done randomly independent of the
number of values associated with each key. As a result, randomly
choosing k keys may not evict the correct amount of bytes, it may
also evict more than the requisite bytes.

This loosens the restriction that eviction always evict exactly
the correct number of bytes or more in exchange for better eviction
performance. In production, this allows us to evict more often without
as large a penalty to customer query performance, which blocks on
eviction when upqueries occur.

Change-Id: I06c3a881ca392cb11aded74252249cfd679e6f89
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/978
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>