ops/cfn: Only check memory every 15 seconds.
Performing the memory check in the readyset server and doing
the subsequent evictions is expensive and can block the execution
of queries during processing. If this occurs at 1 second intervals,
we see a significant drop in query performance. This CL bumps this
up to every 15 seconds to make eviction play nice with the majority
of queries.

Change-Id: Iab2912c32a591d9e572184ea12f133c5d0ce0f17
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/979
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>