ensure-ebs-volume: Support test notification
AWS adds a test notification to the SQS queue on creation. This was
causing the volume script to get stuck in a loop. This handles by adding
parsing support and then always removing the test message.

Also better support parsing the actual message sent on termination and
a test for the lifecycle message pulled from SQS.

Change-Id: I99af9d794a6675ddf44956bd2faec10e29a1726e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/987
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>