metrics: Make NoriaMetricsRecorder thread safe.
Metrics can be recorded from any of the threads in ReadySet.
The metrics-rs crate itself does not guarantee that that recorders
are called in a thread-safe maner. Recorder's themselves must
guarantee this.

https://gerrit.readyset.name/c/readyset/+/957 removed
BufferedMetricsRecorder which enforced this thread-safety
via a RwLock over the recorders. This does not apply to the
PrometheusRecorder which is already thread-safe.

Change-Id: Ic704335cb3b3a1df235ae3d6f85c8d6c5a288f3e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/991
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI