server: Tweak log levels
Adjust the levels of logging throughout the server and dataflow crates,
primarily to bring the relatively verbose logging around migrations down
from info-level to debug-level; these are useful to us when debugging
things but things like a log event at the creation of every new node
shouldn't really be visible to customers at release-level. I've left in
coarser migration things, however, such as the beginning and end of
migration and migration planning, and the beginning and end of replays.

Change-Id: I64f7fa2ecac701832bbabc4980f08a2de33db1cd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/992
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>