ops/cfn: Implement DependsOn for ASG Teardown Scenarios
When CloudFormation attempts to delete an autoscaling group, we need to
be sure we're not tearing down resources that are associated with the
asg. In this CL we'll be trying to pin down those resources.

Additionally, the HeartbeatTimeout on the EC2 LifecycleHook controls how
long the ASG will wait when it fails to send lifecycle transition
messages to SQS. By default, this is set to 1-hour, so in this CL I'm
proposing that we start off by cutting this in half. We may need to
iterate here. You can find the docs on this setting here:
https://go.aws/31iROUR

Once we confirm that this fixes the issue, we can back-port these
changes, as well as the LifecycleHook approach in general into the
Postgres templates.

Change-Id: Id4741bc9bf2f2ebb425c6c1760cb3bcc898d50fd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/994
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>