dataflow: Store a full index in IndexedState
Rather than just storing the key columns on IndexedState, store a full
index - this'll be used later to get a reference to that index when we
evict from a MemoryState. To support this change, this also changes
IndexedState::new to take an owned Index instead of a reference to one,
which threads up through State::add_key and State::add_weak_key,
eventually avoiding several clones, which is nice.

Change-Id: I129c9cddea8413b2aa6fb46dd60d0814e8991c75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1007
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>