ops: Import nighly and default queues into terraform
Import the cloudformation stack resources for the buildkite default and
nighly queues into terraform, leaving the former in place for now but
replacing the latter with an explicit queue named after the size of the
instances it provisions rather than just the word "nightly", updating
the nightly pipeline itself to use this for its agent targeting rules.

I considered briefly doing the same thing with the default queue, but
decided against it as I think the debugging loop of having
steps (especially dynamically uploaded ones) never run unless they
explicitly select an agent targeting rule is too obtuse to make it worth
it.

Change-Id: Ide36c1aea1901d91ef5d359f2352ada15bfafb9e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1012
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI