benchmarks: Add abstraction for benchmark setup and execution.
This adds the `BenchmarkControl` abstraction and a common set of
parameters that benchmarks can easily add to their input parameters,
`BenchmarkParameters`. This will allow running every benchmark
uniformly through calls to `setup` and `benchmark`.

In the future the binary will be extended to run specific
benchmarks from the set of benchmarks in the `Benchmark` set. As
it is implemented now, benchmarks can be run through using a
subcommand with:
`./benchmarks <benchmark name> --<benchmark params>`.

Change-Id: I96f2e9fb4f9ec97f1c61e7cdd828d53509c1a514
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1037
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>