images: Cleanup readyset-monitor images
Since metrics aggregator is something we made specifically for ReadySet
I've moved everything configuring it into a provisioner for the
readyset-monitor image. If we want to add more custom things to it,
it would belong in here.

This ensures the AWS user-data script is only being set as the final
script of any particular image we are building.

The other parts like vector aggregator and such are more "reusable"
portions that we could recombine in other ways.

Change-Id: Idcfecd3997b6410bf9ed7e5d7d261507313c1a47
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1026
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>