ci: Improve build Dockerfile and replace sccache
We were writing the cargo config to ~/.cargo instead of CARGO_HOME which
is set by the Rust container to /usr/local/cargo which seems like it was
not reading from at all.

sccache does not appear to work with our S3 configuration and instead we
need to use cachepot so installing this instead.

Change-Id: I14e707e52dbb501a4d63324619e083c82ea6dcec
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1027
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>