Bump rocksdb to 0.17
Bump the version of rocksdb we're depending on to 0.17, primarily to fix
a build issue with conditional compilation of sse4.2 issues that's
preventing us from using sccache in CI but also generally to get us on
newer versions with lots of bugs fixed.

To get this working and to work around the long-standing dependency
version conflict between bitvec and funty[0], I had to patch bindgen to
the latest commit on their HEAD since that upgrades cexpr, which in turn
upgrades their dependency on nom, which in turn upgrades bitvec and
funty to versions without the conflict. I've asked[1] bindgen to
release a new version with the cexpr version bump, after which I'll make
a PR to rust-rocksdb itself to bump its dependency on bindgen to that
version and we should be able to get out of version-patch land.

[0]: https://github.com/bitvecto-rs/bitvec/issues/105
[1]: https://github.com/rust-lang/rust-bindgen/pull/2107#issuecomment-961279850

Change-Id: If1a84e505d39dfd2efa4833b1f33b057b2b4975e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1030
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>