noria: Split out noria-errors crate
Split out `noria/noria/src/errors.rs` into a new `noria/errors`
crate (named noria-errors), and update all the places where we depend on
it to import from there instead of from noria (with the exception, in a
few places, of `ReadySetError` and `ReadySetResult`, which are both `pub
use`d from noria for now, though I'd like to change that too at some
point in the future).

This primarily paves the way for breaking other stuff out of the noria/
crate too - specifically, I'd like to make DataType and supporting stuff
be defined in its own crate (since lots of stuff depends on noria/ just
for DataType), but since the conversions for DataType return
ReadySetError that requires errors to be in its own crate too.
Regardless, since this is the error type we use for the entire workspace
it makes sense to have it be in its own crate so things can depend on it
alone without having to depend on the whole of noria/

As part of this there was one small change to the actual ReadySetError
definition that was necessary to get this working: in a couple of error
variants we use `LocalNodeIndex` as a field, but since that type is
defined inside of `noria/` we couldn't do that anymore; instead I've
just made it a bare `usize` since that's what `LocalNodeIndex::id`
returns. This feels like a relatively small concession to make here,
especially since we don't actually *use* this field for anything other
than the Display impl of the error later, and the alternative of
creating a new crate just for LocalNodeIndex felt worse to me.

Also, the conversion from channel::tcp::SendError to ReadySetError moved
to the tcp::channel module instead, but there's no loss of functionality
there.

Change-Id: I2a36cf5e54a7a0ded5cf1e1d51bb50413c53d80e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1034
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>