substrate: Add new sccache bucket in build
Currently, we are pointing our builds at a bucket for caching which is
in a separate account and region from where are builds are actually
being run which is likely causing issues with permissions.

This creates a new bucket for the purposes of being used by sccache.

Change-Id: I40968d6bbfddf313f9efe8b328efd70f5913ea92
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1039
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Terraform-Plan-Review: Griffin Smith <griffin@readyset.io>