buildkite: Configure builds to use cachepot.
Use the new bucket and remove the credential URL as it was including the
role which now changes between different queues and probably not needed
anymore.

Also removed some duplicated volume defintions and env variables that
were not being passed through to the docker containers while I was here
going through everything.

Change-Id: I4f941626b8c366d6804f65bd3ae0ad052b6556b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1040
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>