ops/cfn: postgres cf template fine-tuning
Implementing lifecycle hook code from MySQL CF templates and
standardizing the naming on the Postgres side to match the mysql stack's
naming.

Dialing in the configurations in PG to pass around the List<Subnets>.

Since we're enforcing an instance type constraint on the nested stacks,
it makes sense to bubble those constraints up to the super template.
That change is included in this CL.

Also removed the constraints around the monitoring server since we're
not explicitly using EBS optimization there.

Added .taskcat.yml entry to include ReadySet PostgreSQL for testing.

Change-Id: I9223fdc49a4d50fa20538a30a7da17f5b17dbe53
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1044
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>