benchmarks: Add annotations for query parameters.
When generating query parameters for prepared statements, we want
to be able to (1) generate keys that will match rows in the
database, and (2) generate keys that follow customer request
distributions. Since we do not have knowledge of *all* the
rows in a customer's database, we rely on hints that can be
passed into the read benchmark to enable this.

These hints take the form of a `AnnotationSpec`, an annotation for
each parameter that indicate the distribution, min, and max value
of each parameter. For example for a query: SELECT * FROM table
WHERE table.id = ?, the read spec file would be: uniform 0 1000.
This indicates that the query parameter indicated by ? should be
pulled from a uniform distribution between 0 and 1000.

Change-Id: I885b3f0c3729a81aee52349e97a5b62d80470867
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1116
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>