Avoid cloning select statements and region when executing
Currently the select statements are fully cloned every execution
from cache in order to avoid having two mutable references at the
same time. This change reshuffles some parameters so that do_read
no longer takes a mutable reference to self. In addition it no
longer requires a String, but an &str for region.

Change-Id: I440e10211adcb3cde262afa7774023ea4da534d3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1049
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI