noria: Added timeout mechanism when creating a NoriaConnector.
When creating a new `NoriaConnector`, we were previously trying to connect to
Noria without any timeout mechanism, waiting for as long as it takes without
erroring out.
Since the Noria `Leader` is a single-threaded process, we found out that there
are moments where migrations were taking too long (as it can happen), and we
were then unable to create new connections to Noria during that time.
This commit allows us not to hang the process while waiting for new connections,
but two issues still remain:
1. If the timeout occurs, then no connection is made to Noria. That means that
all request are forward to the fallback database. We should have a retry
mechanism in the background to try to connect again.
2. Long-running migrations make the `Leader` unable to process new messages
until they are done. We should run the migrations on a separate thread (and
still run them in order).

Change-Id: Ie02aaf9fc2e3ebaa99fd70af2064d9bbb4f55ca3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1050
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>