noria/server: Use actual allocated memory for eviction
Currently, the in-memory state structures make a vague attempt at
estimating the amount of memory they're using by summing together the
estimated heap usage of the various values that're being inserted. This
is *way* off - both categorically and proportionally - proportionally
because there's a lot of extra on-heap machinery inside of rust hashmaps
besides just the keys and values that're getting inserted (to the tune
of about 8x overhead if we're just filling holes with empty sets of
rows), and categorically because rust vectors (which are used as the
backing storage for hashmaps), like basically all vectors, resize
quadratically, to amortize the cost of copying over the entire contents
of the vector on a resize.

I played around a little bit with trying to figure out a good way
of *actually* estimating the amount of memory used, but there's really
just too much going on inside of hashbrown etc.  to be able to do this
well - not to mention that an update to the rust standard library could
easily result in us being way off on our memory usage tracking again!

Instead of trying to *estimate* the memory used, then, this commit
tracks the *actual* memory allocated to the heap, by calling into
jemalloc's mallctl api by way of the jemalloc_ctl crate.  Since the only
thing we need for deciding *when* to evict is the total heap memory
used, this works perfectly fine - then, we can use the state_sizes
reported by the individual node states to decide *where* to evict,
adjusted proportionally for the measuring error.

Since we're reporting the full heap memory used by the process, and by
default tests all run in the same process, the aggressive_eviction test
needed to be broken out into a separate process by use of the
`rusty_fork` crate, otherwise if it ran alongside other tests it would
evict *so* aggressively that replays would never complete.

Fixes: ENG-726
Change-Id: I4f5a5c0d11f1441112625e5059b778c92d910dc4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1054
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>