Number placeholder literals
Adds a pass in the adapter to replace all placeholder literals with
numbered placeholders. This is to support CL-1553 which maps
placeholder indices to reader keys.

This pass could have been added in the server. However, a thought was
that we would want to manipulate placeholder numbers in the case where
some placeholders were already numbered by the user. This was probably a
bad assumption, and it is likely that we will want to remove this
rewrite from the adapter and add it to sql passes in the server.

Change-Id: Iee76c912924b563e2b4a618e4986aadb6a3da929
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1552
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>