nom-sql: Parse SERIAL and BIGSERIAL
Add new SqlType::Serial and BigSerial types and parsing to nom-sql, and
support both types everywhere we're pattern matching on sql types.

Since SERIAL is just an alias for INT and BIGSERIAL is just an alias
BIGINT for with some extra write-semantics about the default value,
these are implemented as INT4 and INT8 respectively in noria-psql, and
DataType::Int and DataType::BigInt respectively in noria. Note
especially that in both cases these are *not* unsigned, since even
though a negative value will never get *generated* for a serial column,
it is still possible to explicitly *insert* a serial value into a serial
column.

Change-Id: I5a133d149a0da78a59a64bfad26b80e6aae6d164
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1059
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>