noria-mysql: Remove panics in convert_column
Convert the unimplemented!() panics in
noria_mysql::schema::convert_column to unsupported! error returns, to
avoid panics.  Since this is happening at the *very* edge of the
protocol, extra care has to be taken to make this bubble up to the
client as a proper error packet, rather than just being returned from
the function (which would cause the connection to drop).

Change-Id: Id4e41b733a62204f8c33a2427ec5e296c74b1635
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1060
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>