ops/cfn: Implement MySQL air gap support and VPC endpoints
This CL is intended to enable us to run our MySQL stack as an air gapped
environment. Several things needed to happen to get this working, as
we're leveraging several different APIs: ec2, sqs, autoscaling, and
cloudformation (aws-cfn-bootstrap).

Now that we're implementing more network infrastructure on-top of the
quick-start VPC, it made sense to abstract and centralize these
components into a designated stack. As a benefit, we're able to reduce
complexity in the current Postgres and MySQL stacks.

Speaking of Postgres, I'll submit a follow up PR to dial that stack in.
It should come in shortly after this CL; just wanting to get the MySQL
bits out ASAP for IRL demo.

Change-Id: If0941bda49bfdbc7930028410a6821eccdecd4c0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1066
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>