Stop considering self in check of running instances
We found a bug in ensure-ebs-volume where when filtering for running
instances, we considered ourselves from the retrieved list. This meant
that we believed we needed to wait for ourselves to shut down before
proceeding to acquire the ebs volume, ensuring it never got acquired.

This fixes this behavior by adding an additional check that we do not
consider ourselves from the retrieved instances list.

Change-Id: I6ddaa0889481d059ae3c5abf8d23a3d8f00704a4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1067
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>