adapter: Don't extend recipe in request path when live qca is enabled.
When live QCA is enabled, the query reconciler handles performing
all migrations. The request path does not need to extend the
noria recipe to retrieve the view, since the view should already
be created. This removes the redundent `extend_recipe` call from
the request path and simply checks that the view we require exists.
This will prevent the spam of `extend_recipe` that occurs when
a view is first created and queries related to it first start
hitting Noria.

Change-Id: I354f4ca0fe8f1be14dd44869d937d1b7c427a468
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1072
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>