ops/cfn: supplemental IAM policies needed for EC2 VPCE interface
After closer inspection of the ensure EBS volume logs, it would appear
that we're hitting some snags whilst attempting to execute
ec2:AttachVolume commands and the like.

The goal of this CL is to extend the EC2 VPC endpoint's IAM policy to
unlock this capability. Currently, server nodes aren't able to attach
additional volumes which could present a problem.

Change-Id: I94b5c7d2a369afb6f457ba3a9fb3226b66c73833
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1073
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>