controller: Replace assert with error on spurious aliasing.
When two queries have different schemas but the same name, we
currently panic due to an assertion in the recipe code. This
replaces those assertions with errors so we can gracefully fail
when this occurs.

This issue arose during the IRL demo where two queries in a
mysqldump included created views with the same name but
different schemas. When the binlog replicator extended recipe's
associated with this DDL, the noria-server instance panics.

Change-Id: I0b60986e113f0624171b73b88265e1437622ad25
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1074
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>