query reconciler: Add unsupported state for queries we won't retry.
We currently do not support certain queries, for example, topk.
Retrying these queries against Noria repeatedly in the query
reconciler will never yield a different result in the time
allotted to retry a query. As a result, when we received specific
errors from Noria as a result of preparing the statement, we should
assign the query the new `Unsupported` state and never retry it.

Notably, this does not include when tables do not exist on attempt
to query as a schema change may not have propagated from the backend.

This CL also expands error messages in the query reconciler to
include the error and query it failed on.

Change-Id: I0819aa14538d901af0fcca3ee0494068174b74f8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1076
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI